Control.Concurrent.Speculation.Class

Description

Versions of the combinators from the speculation package with the signature rearranged to enable them to be used directly as actions in the Cont and ContT monads or any other Codensity-shaped monad.

Documentation

class MonadSpec m where

Methods

specByM :: (a -> a -> Bool) -> a -> a -> m a

spec with a user supplied comparison function

specM :: (MonadSpec m, Eq a) => a -> a -> m a

When a is unevaluated, spec g a evaluates the current continuation with g while testing if g == a, if they differ, it re-evalutes the continuation with a. If a was already evaluated, the continuation is just directly applied to a instead.

specOnM :: (MonadSpec m, Eq c) => (a -> c) -> a -> a -> m a

spec' with a user supplied comparison function

Basic speculation